+++ vuapos.Presentation/Views/Promotion/PromotionPage.xaml.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices.WindowsRuntime;
+using System.Threading.Tasks;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Controls.Primitives;
@@ -12,6 +14,7 @@ using Microsoft.UI.Xaml.Media;
 using Microsoft.UI.Xaml.Navigation;
 using vuapos.Presentation.Models;
 using vuapos.Presentation.Services;
+using vuapos.Presentation.Services.Interfaces;
 using vuapos.Presentation.Views.Product;
 using Windows.Foundation;
 using Windows.Foundation.Collections;
@@ -24,11 +27,14 @@ namespace vuapos.Presentation.Views.Promotion
     public sealed partial class PromotionPage : UserControl
     {
         public PromotionViewModel ViewModel { get; }
+        private readonly IUserSession _userSession;
 
         public PromotionPage()
         {
             this.InitializeComponent();
             ViewModel = new PromotionViewModel();
+            _userSession = App.Services.GetRequiredService<IUserSession>();
+
             LoadInitialData();
         }
 
@@ -42,10 +48,12 @@ namespace vuapos.Presentation.Views.Promotion
         {
             try
             {
-               var promotionDialog = new AddPromotionDialog(ViewModel) 
-               {
-                   XamlRoot = this.XamlRoot
-               };
+                if (_userSession.role != "MANAGER")
+                    throw new Exception("You do not have permission to add products");
+                var promotionDialog = new AddPromotionDialog(ViewModel) 
+                   {
+                       XamlRoot = this.XamlRoot
+                   };
                 await promotionDialog.ShowAsync();
                 ViewModel.currentPage = 1;
                 await ViewModel.LoadPromotionsAsync();
@@ -53,13 +61,7 @@ namespace vuapos.Presentation.Views.Promotion
             }
             catch (Exception ex)
             {
-                await new ContentDialog
-                {
-                    Title = "Error",
-                    Content = $"Failed to open dialog: {ex.Message}",
-                    CloseButtonText = "OK",
-                    XamlRoot = this.XamlRoot
-                }.ShowAsync();
+                await ShowErrorDialogAsync($"Failed to open dialog: {ex.Message}");
             }
         }
 
@@ -85,51 +87,70 @@ namespace vuapos.Presentation.Views.Promotion
 
         }
 
+        private async Task ShowErrorDialogAsync(string message)
+        {
+            await new ContentDialog
+            {
+                Title = "Error",
+                Content = message,
+                CloseButtonText = "OK",
+                XamlRoot = this.XamlRoot
+            }.ShowAsync();
+        }
         private async void EditPromotion_Click(object sender, RoutedEventArgs e)
         {
-            if (sender is Button button && button.DataContext is Promotion promotion)
+            try
             {
-                var editPromotionDialog = new EditPromotionDialog(ViewModel, promotion)
+                if (_userSession.role != "MANAGER")
+                    throw new Exception("You do not have permission to add products");
+                if (sender is Button button && button.DataContext is Promotion promotion)
                 {
-                    XamlRoot = this.XamlRoot
-                };
-                await editPromotionDialog.ShowAsync();
-                await ViewModel.LoadPromotionsAsync();
+                    var editPromotionDialog = new EditPromotionDialog(ViewModel, promotion)
+                    {
+                        XamlRoot = this.XamlRoot
+                    };
+                    await editPromotionDialog.ShowAsync();
+                    await ViewModel.LoadPromotionsAsync();
+                }
             }
-            else
+            catch(Exception ex)
             {
-                await new ContentDialog
-                {
-                    Title = "Error",
-                    Content = "Failed to open dialog.",
-                    CloseButtonText = "OK",
-                    XamlRoot = this.XamlRoot
-                }.ShowAsync();
+                await ShowErrorDialogAsync($"Failed to open dialog: {ex.Message}");
             }
         }
         private async void DeletePromotion_Click(object sender, RoutedEventArgs e)
         {
-            if (sender is Button button && button.DataContext is Promotion promotion)
+            try
             {
-                var result = await new ContentDialog
+                if (_userSession.role != "MANAGER")
+                    throw new Exception("You do not have permission to add products");
+                if (sender is Button button && button.DataContext is Promotion promotion)
                 {
-                    Title = "Delete Category",
-                    Content = $"Are you sure you want to delete {promotion.Name}?",
-                    PrimaryButtonText = "Delete",
-                    CloseButtonText = "Cancel",
-                    XamlRoot = this.XamlRoot
-                }.ShowAsync();
-                if (result == ContentDialogResult.Primary)
-                {
-                    await ViewModel.DeletePromotionAsync(promotion.Promotion_Id);
-                    //await ViewModel.LoadPromotionsAsync();
-                    ViewModel.currentPage = 1;
-                    await ViewModel.LoadPromotionsAsync();
-                    CurrentPageTextBlock.Text = $"Page {ViewModel.currentPage} of {ViewModel.totalPages}";
-                    
+                    var result = await new ContentDialog
+                    {
+                        Title = "Delete Category",
+                        Content = $"Are you sure you want to delete {promotion.Name}?",
+                        PrimaryButtonText = "Delete",
+                        CloseButtonText = "Cancel",
+                        XamlRoot = this.XamlRoot
+                    }.ShowAsync();
+                    if (result == ContentDialogResult.Primary)
+                    {
+                        await ViewModel.DeletePromotionAsync(promotion.Promotion_Id);
+                        //await ViewModel.LoadPromotionsAsync();
+                        ViewModel.currentPage = 1;
+                        await ViewModel.LoadPromotionsAsync();
+                        CurrentPageTextBlock.Text = $"Page {ViewModel.currentPage} of {ViewModel.totalPages}";
 
+
+                    }
                 }
             }
+            catch (Exception ex)
+            {
+                await ShowErrorDialogAsync($"Failed to open dialog: {ex.Message}");
+            }
+            
         }
 
         private async void PromotionPage_Loaded(object sender, RoutedEventArgs e)
